refactor cypher mcp
