+++ servers/mcp-neo4j-cypher/inspector.sh
@@ -1,0 +1,1 @@
+npx @modelcontextprotocol/inspector uv run mcp dev src/mcp_neo4j_cypher/server.py --log-level debug
