+++ servers/mcp-neo4j-cypher/pyproject.toml
@@ -4,7 +4,11 @@ version = "0.1.1"
 description = "A simple Neo4j MCP server"
 readme = "README.md"
 requires-python = ">=3.10"
-dependencies = ["mcp>=1.6.0", "neo4j>=5.26.0"]
+dependencies = [
+    "mcp[cli]>=1.6.0",
+    "neo4j>=5.26.0",
+    "pydantic>=2.10.1",
+]
 
 [build-system]
 requires = ["hatchling"]
