+++ servers/mcp-neo4j-cypher/pyrightconfig.json
@@ -1,0 +1,4 @@
+{
+    "venvPath": ".",
+    "venv": ".venv"
+}
